Pull-To-Refresh, Snackbar, NetworkHelper

+++ app/src/main/java/com/vladbytsyuk/weatherforecastproject/view/fragments/WeatherForecastFragment.java
@@ -2,14 +2,12 @@ package com.vladbytsyuk.weatherforecastproject.view.fragments;
 
 import android.app.Fragment;
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.util.Log;
+import android.support.v4.widget.SwipeRefreshLayout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Adapter;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.ProgressBar;
@@ -17,33 +15,49 @@ import android.widget.TextView;
 
 import com.vladbytsyuk.weatherforecastproject.controller.DBManager;
 import com.vladbytsyuk.weatherforecastproject.R;
+import com.vladbytsyuk.weatherforecastproject.controller.NetworkHelper;
 import com.vladbytsyuk.weatherforecastproject.controller.adapters.WeatherForecastAdapter;
 import com.vladbytsyuk.weatherforecastproject.model.WeatherForecast;
+import com.vladbytsyuk.weatherforecastproject.controller.PreferencesHelper;
 
 /**
  * Created by VladBytsyuk on 22.11.2015.
  */
-public class WeatherForecastFragment extends Fragment {
+public class WeatherForecastFragment extends Fragment
+        implements SwipeRefreshLayout.OnRefreshListener {
     private Context context;
     private DBManager dbManager;
     private TextView textViewCity;
     private ListView listView;
-    private ProgressBar progressBar;
     private WeatherForecastAdapter adapter;
+    private SwipeRefreshLayout swipeRefreshLayout;
 
     public interface OnItemPressed {
         void itemPressed(WeatherForecast weatherForecast);
     }
-    private OnItemPressed listener_item_pressed;
+    private OnItemPressed listenerItemPressed;
 
-    public void setOnItemPressedListener(OnItemPressed listener_item_pressed) {
-        this.listener_item_pressed = listener_item_pressed;
+    public void setOnItemPressedListener(OnItemPressed listenerItemPressed) {
+        this.listenerItemPressed = listenerItemPressed;
     }
 
     public void removeOnItemPressedListener() {
-        this.listener_item_pressed = null;
+        this.listenerItemPressed = null;
     }
 
+
+    public interface OnRefresh {
+        void showNetworkIssuesSnackbar();
+    }
+    private OnRefresh listenerRefresh;
+
+    public void setOnRefreshListener(OnRefresh listenerRefresh) {
+        this.listenerRefresh = listenerRefresh;
+    }
+
+    public void removeOnRefreshListener() {
+        this.listenerRefresh = null;
+    }
 
     @Nullable
     @Override
@@ -55,47 +69,54 @@ public class WeatherForecastFragment extends Fragment {
         dbManager = new DBManager(context);
         if (dbManager.isDBEmpty())
             dbManager.downloadWeather();
-        textViewCity = (TextView) rootView.findViewById(R.id.textViewWeatherForecastCity);
-        textViewCity.setText(getPreference(R.string.city));
 
-        listView = (ListView) rootView.findViewById(R.id.listViewWeatherForecast);
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                if (listener_item_pressed != null)
-                    listener_item_pressed.itemPressed(adapter.getItem(position));
-            }
-        });
-        adapter = new WeatherForecastAdapter(context, dbManager.getWeatherForecasts());
-        listView.setAdapter(adapter);
+        swipeRefreshLayout = (SwipeRefreshLayout) rootView.findViewById(R.id.swipeRefreshLayoutWeatherForecast);
+        swipeRefreshLayout.setOnRefreshListener(this);
+        swipeRefreshLayout.setColorSchemeResources(android.R.color.holo_blue_light,
+                android.R.color.holo_blue_bright,
+                android.R.color.holo_blue_dark,
+                android.R.color.holo_blue_bright);
 
-        progressBar = (ProgressBar) rootView.findViewById(R.id.progressBarWeatherForecast);
-        progressBar.setVisibility(View.INVISIBLE);
+        rootView = initListView(rootView);
+        rootView = initTextView(rootView);
 
         return rootView;
     }
 
-    private String getPreference(int preference) {
-        SharedPreferences settings = getActivity()
-                .getSharedPreferences(getActivity().getString(R.string.shared_preferences_file_name), Context.MODE_PRIVATE);
-        return settings.getString(getActivity().getString(preference), null);
+    private View initListView(View rootView) {
+        listView = (ListView) rootView.findViewById(R.id.listViewWeatherForecast);
+        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                if (listenerItemPressed != null)
+                    listenerItemPressed.itemPressed(adapter.getItem(position));
+            }
+        });
+        adapter = new WeatherForecastAdapter(context, dbManager.getWeatherForecasts());
+        listView.setAdapter(adapter);
+        return rootView;
     }
 
+    private View initTextView(View rootView) {
+        textViewCity = (TextView) rootView.findViewById(R.id.textViewWeatherForecastCity);
+        textViewCity.setText(PreferencesHelper.getPreference(R.string.city));
+        return rootView;
+    }
 
     public void refresh() {
-        hideListView();
-        dbManager.drop();
-        dbManager.downloadWeather();
-        showListView();
+        if (NetworkHelper.isNetworkAvailable()) {
+            dbManager.drop();
+            dbManager.downloadWeather();
+            adapter = new WeatherForecastAdapter(context, dbManager.getWeatherForecasts());
+            listView.setAdapter(adapter);
+        } else  if (listenerRefresh != null) {
+            listenerRefresh.showNetworkIssuesSnackbar();
+        }
     }
 
-    private void hideListView() {
-        listView.setVisibility(View.INVISIBLE);
-        progressBar.setVisibility(View.VISIBLE);
-    }
-
-    private void showListView() {
-        progressBar.setVisibility(View.INVISIBLE);
-        listView.setVisibility(View.VISIBLE);
+    @Override
+    public void onRefresh() {
+        refresh();
+        swipeRefreshLayout.setRefreshing(false);
     }
 }
